Fix check for finding PkgConfig
This is a follow up of !1354. There PkgConfig is only looked up if
the `pkg_check_modules` macro hasn't been define. The problem is that
the scope of macros seem to be global, and finding PkgConfig in another
directory defines that macro. However, variable scope is per directory
and the cmake of PkgConfig doesn't really work.

This changes the check to use the `PKG_CONFIG_FOUND` variable