Fix TProfile creation in ROOT monitoring sink when SetDefaultSumw2 is used
When the static function TH1::SetDefaultSumw2 had been called (passing true), `Sumw2(true)` is automatically
called by the constructor of TProfile. This is a problem because in the profiles in Gaudi::Accumulators we do
not keep track of the sum of squares of weights and consequently don't set fBinSumw2 of the TProfile, which in
turn leads to a self-inconsistent TProfile object. The "fix" is to disable sum of squares explicitly.