Add CMake function gaudi_add_pytest
gaudi_add_pytest is used to declare pytest tests. Each of the test
files discovered by pytest are mapped to one ctest test, to be able to
run them in parallel.

Tests are collected when ctest is invoked and/or at build time, and only
in case changes are detected in scanned files.