New version of PartProp service
This is a port of the LHCb PartProp service into Gaudi.

Namespace for the service is updated to be part of Gaudi.
Various examples and tests are all ported.
Copyright and license are updatded to match Gaudi.
Test algorithm is converted to a plain Algorithm
and not a GaudiAlgorithm.
Considerable clean-up of Python parts to improve flake8
and pylint compliance. (There are some import statements
that have side effects, which are excepted from tests,
F401.)