Add support for unordered set properties
Change the Python type of `std::unordered_set` properties from `list` to `set`.
Parsers only required minor updates and special treatment of the empty
set, which can only be represented as `set()` in Python (`{}` is
reserved for dictionaries).

This will certainly break client code, i.e.
```
   MyAlg.ExtraInputs += ['foo']
```
will have to be updated to
```
  MyAlg.ExtraInputs.add('foo')
```

For the moment only `std::unordered_set` is mapped to a Python `set`.
Regular `std::set` remains as Python `list` to maintain insertion order.