Implemented Root like histograms
Their usage is similar to standard histograms, names are prefixed with "Root" and at this stage only non profile, non weighted versions are provided, for dimensions 1,  and 3 only.
The main difference with standard histograms is that they compute statistical data on the input on top of the strict histogram, in particular they keep sum and sum of squares of all variables so that mean and standard_deviation of the original data can be computed.
This behavior mimics Root histograms, which allows to convert the new histograms to Root ones and have the expected behavior, while the Gaudi histograms had no statistical data and their conversion to Root was somehow recomputing them (Root did) based on the histogram content. Obviously the values obtained were not correct.