Improve pytest coverage report modules selection
`pytest` discovers coverage options in all parent directories, so if run the tests from a
subdirectory of a Python project the sets `--cov` we get spurious results in the
coverage reports. The option `--cov-reset` cleans the list so that we correctly
select the modules we want to check.

See https://github.com/pytest-dev/pytest-cov/issues/460 for more details.