DataObjID: on-demand lookup of className
Retrieving the class name to a CLID requires querying the `ClassIDSvc`,
which (at least in the ATLAS implementation) requires taking a lock
during the lookup and in general might be a costly operation. Instead of
doing this in the construction of `DataObjID` only lookup the
class name when needed and then cache the result.

Also cleanup the needlessly complicated code of one-time retrieval of
ClassIDSvc by using a simple static.

This fixes some performance problems in multi-threaded ATLAS
jobs (ATEAM-1054).