Modernized and evolved Auditors
Main goal was to make EventContext available in Auditors to be able to implement thread safe behaviors (e.g. Timingauditor)

On the way, the IAuditor interface was simplified :
  - dropped INamedInterface support
  - dropped StandardEventType support
Both were considered premature optimization at this stage, and the interface was thus heavy (4 methods for before, 4 for after compared to 1 and 1 now)

Old, legagy auditors have been kept and the change is (almost fully) backward compatible.
The new interface is Gaudi::IAuditor and the new base class Gaudi::Auditor. Previous ::IAuditor and ::Auditor are still supported at this stage and Auditors have not yet been migrated.
The old interface is now inheriting from the new one, so that code in Algorithm/Service/AlgTool could be adapted to the new interface, and thus cleanup should be easy in the future.
Note that there is no equivalent of ::CommonAuditor anymore. The new interface basically is it.