Refactoring of (I)AlgExecStateSvc to hide internal types and optimize implementation
Practically :
  - dropped public type AlgStateMap_t
  - dropped accessor to internal type
  - replaced the original std::map with a custom type using an unordered_map combined with a vector to be more efficient when running full scans

The main user facing change is a change of API in IAlgExecStateSvc where algExecState does not return anymore a reference to an AlgExecState but a plain object which is actually a wrapper to an AlgExecState.
This wrapper, still called AlgExecState (while the regular state is now AlgExecStateSvc::AlgExecStateInternal) keep references to the service, the context and an identifier of the algorithm called AlgKet (practically a size_t).
This identifier has been added to the Algorithm class (not touching the IAlgorithm interface) and it used for efficient access to ExecState.
Current implementation is that this is the index in the underlying vector of states.
The AlgKey can also be retrieved when addAlg is called, allowing to not even use the algExecState interface in some cases.