Make IInterface::cast try queryInterface if i_cast fails
This is meant as a temporary measure while we ensure all special
queryInterface implementations are replaced with equivalent i_cast
implementations.

To be able to use queryInterface from IInterface::cast, I added the
special method IInterface::decRef, that reduces the reference count
without deleting the object if the ref count goes to 0.