Fixed behavior of Streamers of ROOT in case EnableThreadSafety is activated
Before this fix, it would crash as we override streamers for SmartRef in particular using IOHandler and that one was not implementing the Generate method.
This lead to the Generate method of the base class to be called and potentially to seg fault if I/O was done outside the main thread.

A new test was added to exercise that case.