Added a SimpleWriter class in NTuple/Writer.h
This one allows free creation of extra entries in the NTuple without the restriction of one entry per event.
It also does not impose to go through the transform method systematically.

The Writer class remains unchanged from the user point of view and is based on the new SimpleWriter.

Also the WriterMixin class was merged into SimpleWriter.