Make CUDA tests optional
Build the CUDA code by default on a regular node as part of the regular
LCG view build. The separate "GPU" test can be triggered manually to
build/run on a GPU node.

Also use nvidia-smi to check if there is a GPU installed and
enable/disable GPU tests accordingly.