Fixed string representation in option files (json in particular)
repr should have been used so that escape sequences are properly handled.
This was so far not possible as it was creating single quoted strings not understood by the PropertySvc (by the StringGrammar to be precise).
This has now been fixed in https://gitlab.cern.ch/gaudi/Gaudi/-/merge_requests/1839 allowing to fix all cases.

A Test has been added, with a previously problematic case (an r string containing "\.")